Add many improvments and fixes to trace decompression routine
We now consider two subsequent packets to make correct decision
whether a packet of format 0, 1 or 2 refers to an address (as
synchronization point) or an unpredictable discontinuity. To make that
effective we currently forbid arbitrary "resync" packets. The routine
was also cleaned up a bit by moving common parts into static
functions. The comments explain the changes in a more detailed
fashion. prefetch() is just a fragment of the original linux source
code, intended for processor specific optimizations which we don't
care about. There are still a few logging statements in the code which
will later on be disabled with compile time switches.