Emit packet after F_SYNC exception packet to jump over vec table
The trap vector table is being modified during runtime, which resembles
self-modifying code. Since our decoder has no runtime information about the
state of the program code we add this quick hack which, by emitting an aditional
packet, allows us to know the destination of the jump instruction in the vector
table. Note that this assumes that the entries of that table are always jump
instructions.