Replace linux/list.h with sys/queue.h
We do this to make a further transition to bsd-style licensed code easier. The
semantics of both list implementations are about the same. Furthermore we now
always iterate packet and instruction lists in the forward fashion and append
new packets at the end of the queue instead of doing the opposite. It seems to
be more natural this way to me. The only real difference is that list heads are
now seperate for each struct (struct trdb_packet_head and struct trdb_instr_head
instead of the generic struct list_head from list.h.) and that those have to
intitialized explicitely with TAILQ_INIT. Newer version of queue.h don't have
this problem but are not available on linux. Maybe its will be easier if we just
copy the newer queue.h from openbsd instead of using the system provider header
file.