Enable -Wextra and fix the new warnings
We disable -Wno-missing-field-initializers because it seems to be mislead by
e.g. omitted fields in struct initialization (which should be set to zero
according to the standard).

Many warnings are just related to unused variables, where appropriate we
silence them by doing a (void)var. In other cases we use the empty main() function
when we don't need its arguments.

trdb_disassemble_trace_with_bfd() is now properly declared.

We remove any hacks pertaining initialization ({{0}}). It is mistake of the
compiler to warn when we initialize with {0}. Let's just ignore it.

Add /* fall through */ statements to indicate that we really want it.