Remove modports in interface for tb
APB_BUS is already a modport and it is not possible to nest modports (or
interfaces). So we get rid of that. Generally it is also recommended to
not use modports in testbenches. Interestingly enough that seemed to
have worked just fine with vsim-10.5c but newer version now refuse to
compile.