Accept compressed instruction at tracer interface
This is a change to bring the rtl up-to-date with libtrdb. RI5CY always
provided us with pre-decompressed instructions, but to process traces
produced by spike we have to handle compressed instructions directly.
For that we add new constants generated by riscv-opcodes and check for
compressed branches and jals.

include/trdb_pkg:
Add new constants for compressed instructions.

rtl/trace_debugger:
Detect BEQZ and BNEZ as branches. Handle c_jalr and c_jr as
unpredictable distcontinuity.