Add test target
This allows us to easily feed all the cvs files through our testbench.
Do something like make -j8 test for parallel testing. The wlf changes
are just there to make vsim not crash in that situation.

The other changes are just fixes to make sure that the simulation
results are correctly displayed and that we don't generate a debugdb
when running batch tests.