Add support for 64-bit bus width of output
Two new parameters control the width of the output bus and its address
width. The latter is given by the architecture of the used RISC-V core
while the former is customizable but only 32 and 64 are supported.

The aligner is extended in a straightforward fashion by just replacing
hardcoded values with appropriate parameters. I couldn't think (so far)
of a better way to parametrize the compuation of residual_d and
data_d (except for using an array and indexing into it with a for loop).